next-action-generator/pyproject.toml
[project]
name = "next-action-generator"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "strands-agents>=1.15.0",
    "strands-agents-tools[agent_core_code_interpreter]>=0.2.14",
    "bedrock-agentcore==0.1.0",
    "boto3>=1.40.22",
    "anyio>=4.0.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "aws-opentelemetry-distro>=0.12.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
